requirements-dev.txt
-r requirements-runtime.txt
-r requirements-nightly.txt
pytest==8.3.3
pytest-asyncio==0.24.0
moto==5.0.16
ruff==0.6.9
black==24.8.0
mypy==1.11.2
python-dotenv==1.0.1
mcp==1.1.0
dspy-ai==3.1.3
